refactoring, methods moved to own controller and service
